Start modules refactor
